Created Warehouse services

+++ Backend/backend/src/main/java/com/courier/repository/WarehouseRepository.java
@@ -3,5 +3,5 @@ package com.courier.repository;
 import com.courier.entities.Warehouse;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WarehouseRepository extends JpaRepository<Warehouse, Integer> {
+public interface WarehouseRepository extends JpaRepository<Warehouse, Long> {
 }
